refactor: SelectJdbcTemplate를 JdbcTemplate에 통합
