+++ app/src/main/java/com/techcourse/dao/jdbc/template/JdbcTemplate.java
@@ -2,6 +2,7 @@ package com.techcourse.dao.jdbc.template;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import javax.sql.DataSource;
 import org.slf4j.Logger;
@@ -33,4 +34,36 @@ public abstract class JdbcTemplate {
     protected abstract DataSource getDataSource();
 
     protected abstract void setValues(PreparedStatement pstmt) throws SQLException;
+
+    public Object query() {
+        final String sql = createQuery();
+
+        final DataSource dataSource = this.getDataSource();
+        ResultSet rs = null;
+        try (final Connection conn = dataSource.getConnection();
+            final PreparedStatement pstmt = conn.prepareStatement(sql)) {
+
+            setValues(pstmt);
+            rs = executeQuery(pstmt);
+            LOG.debug("query : {}", sql);
+            return mapRow(rs);
+
+        } catch (SQLException e) {
+            LOG.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (rs != null) {
+                    rs.close();
+                }
+            } catch (SQLException ignored) {
+            }
+        }
+    }
+
+    private ResultSet executeQuery(PreparedStatement pstmt) throws SQLException {
+        return pstmt.executeQuery();
+    }
+
+    protected abstract Object mapRow(ResultSet rs) throws SQLException;
 }
